@@ -6,12 +6,20 @@ import { ROUTES } from '../constants/routes';
 
 export const LessonForm: React.FC = () => {
   const [formData, setFormData] = useState<LessonFormData>({
+    storeId: '',
+    userId: '',
     customerId: '',
-    instructorId: '',
-    date: '',
-    startTime: '',
-    endTime: '',
-    notes: '',
+    postureGroupId: null,
+    condition: '',
+    weight: null,
+    meal: '',
+    memo: '',
+    startDate: '',
+    endDate: '',
+    nextDate: null,
+    nextStoreId: null,
+    nextUserId: null,
+    trainings: [],
   });
   const [loading, setLoading] = useState(false);
   const navigate = useNavigate();
@@ -21,7 +29,7 @@ export const LessonForm: React.FC = () => {
     setLoading(true);
     try {
       await lessonApi.create(formData);
-      navigate(ROUTES.LESSON_HISTORY);
+      navigate(ROUTES.CUSTOMER_SELECT);
     } catch (error) {
       console.error('Error creating lesson:', error);
     } finally {
@@ -30,7 +38,7 @@ export const LessonForm: React.FC = () => {
   };
 
   return (
-    <div>
+    <div className="p-8">
       <h1 className="text-2xl font-bold mb-4">レッスン登録</h1>
       <form onSubmit={handleSubmit} className="space-y-4 max-w-md">
         <div>
@@ -44,50 +52,58 @@ export const LessonForm: React.FC = () => {
           />
         </div>
         <div>
-          <label className="block text-sm font-medium">インストラクターID</label>
+          <label className="block text-sm font-medium">店舗ID</label>
           <input
             type="text"
-            value={formData.instructorId}
-            onChange={(e) => setFormData({ ...formData, instructorId: e.target.value })}
+            value={formData.storeId}
+            onChange={(e) => setFormData({ ...formData, storeId: e.target.value })}
             required
             className="mt-1 block w-full px-3 py-2 border rounded-md"
           />
         </div>
         <div>
-          <label className="block text-sm font-medium">日付</label>
+          <label className="block text-sm font-medium">ユーザーID</label>
+          <input
+            type="text"
+            value={formData.userId}
+            onChange={(e) => setFormData({ ...formData, userId: e.target.value })}
+            required
+            className="mt-1 block w-full px-3 py-2 border rounded-md"
+          />
+        </div>
+        <div>
+          <label className="block text-sm font-medium">開始日</label>
           <input
             type="date"
-            value={formData.date}
-            onChange={(e) => setFormData({ ...formData, date: e.target.value })}
-            required
+            value={formData.startDate || ''}
+            onChange={(e) => setFormData({ ...formData, startDate: e.target.value })}
             className="mt-1 block w-full px-3 py-2 border rounded-md"
           />
         </div>
         <div>
-          <label className="block text-sm font-medium">開始時間</label>
+          <label className="block text-sm font-medium">終了日</label>
           <input
-            type="time"
-            value={formData.startTime}
-            onChange={(e) => setFormData({ ...formData, startTime: e.target.value })}
-            required
+            type="date"
+            value={formData.endDate || ''}
+            onChange={(e) => setFormData({ ...formData, endDate: e.target.value })}
             className="mt-1 block w-full px-3 py-2 border rounded-md"
           />
         </div>
         <div>
-          <label className="block text-sm font-medium">終了時間</label>
+          <label className="block text-sm font-medium">体重 (kg)</label>
           <input
-            type="time"
-            value={formData.endTime}
-            onChange={(e) => setFormData({ ...formData, endTime: e.target.value })}
-            required
+            type="number"
+            step="0.1"
+            value={formData.weight || ''}
+            onChange={(e) => setFormData({ ...formData, weight: e.target.value ? Number(e.target.value) : null })}
             className="mt-1 block w-full px-3 py-2 border rounded-md"
           />
         </div>
         <div>
-          <label className="block text-sm font-medium">備考</label>
+          <label className="block text-sm font-medium">メモ</label>
           <textarea
-            value={formData.notes}
-            onChange={(e) => setFormData({ ...formData, notes: e.target.value })}
+            value={formData.memo || ''}
+            onChange={(e) => setFormData({ ...formData, memo: e.target.value })}
             className="mt-1 block w-full px-3 py-2 border rounded-md"
           />
         </div>
@@ -102,4 +118,3 @@ export const LessonForm: React.FC = () => {
     </div>
   );
 };
-
