@@ -5,13 +5,11 @@ import { MainLayout } from './components/common/MainLayout';
 import { Loading } from './components/common/Loading';
 import { ROUTES } from './constants/routes';
 import { Login } from './pages/Login';
-import { ShopManagement } from './pages/ShopManagement';
 import { CustomerSelect } from './pages/CustomerSelect';
 import { LessonForm } from './pages/LessonForm';
 import { LessonHistory } from './pages/LessonHistory';
 import { CustomerProfile } from './pages/CustomerProfile';
 import { PostureList } from './pages/PostureList';
-import { PostureDetail } from './pages/PostureDetail';
 import { PostureCompare } from './pages/PostureCompare';
 import { CustomerManagement } from './pages/CustomerManagement';
 import { CustomerList } from './pages/CustomerList';
@@ -19,34 +17,19 @@ import { UserManagement } from './pages/UserManagement';
 import { UserList } from './pages/UserList';
 import { HiHome, HiClock } from 'react-icons/hi';
 
-// アイコンコンポーネント
-const HomeIcon = (props: { className?: string }) => {
-  const Icon = HiHome as any;
-  return <Icon {...props} />;
-};
-
-const ClockIcon = (props: { className?: string }) => {
-  const Icon = HiClock as any;
-  return <Icon {...props} />;
-};
-
 // ページごとのメニュー項目
-const shopManagementMenuItems = [
-  { path: ROUTES.SHOP_MANAGEMENT, label: 'Home', icon: <HomeIcon className="w-5 h-5" /> },
-];
-
 const customerSelectMenuItems = [
-  { path: ROUTES.CUSTOMER_SELECT, label: 'Home', icon: <HomeIcon className="w-5 h-5" /> },
+  { path: ROUTES.CUSTOMER_SELECT, label: 'Home', icon: <HiHome className="w-5 h-5" /> },
 ];
 
 // lessonHistoryMenuItemsは動的に生成する必要があるため、
 // Routeコンポーネント内で顧客IDを取得して使用します
 const getLessonHistoryMenuItems = (customerId: string) => [
-  { path: ROUTES.CUSTOMER_SELECT, label: 'Home', icon: <HomeIcon className="w-5 h-5" /> },
+  { path: ROUTES.CUSTOMER_SELECT, label: 'Home', icon: <HiHome className="w-5 h-5" /> },
   {
     path: ROUTES.LESSON_HISTORY.replace(':id', customerId),
     label: '履歴',
-    icon: <ClockIcon className="w-5 h-5" />,
+    icon: <HiClock className="w-5 h-5" />,
     subItems: [
       { path: ROUTES.CUSTOMER_PROFILE.replace(':id', customerId), label: 'プロフィール' },
       { path: ROUTES.POSTURE_LIST.replace(':id', customerId), label: '画像一覧' },
@@ -55,7 +38,7 @@ const getLessonHistoryMenuItems = (customerId: string) => [
 ];
 
 const defaultMenuItems = [
-  { path: ROUTES.SHOP_MANAGEMENT, label: 'Home', icon: <HomeIcon className="w-5 h-5" /> },
+  { path: ROUTES.CUSTOMER_SELECT, label: 'Home', icon: <HiHome className="w-5 h-5" /> },
 ];
 
 const PrivateRoute: React.FC<{ children: React.ReactElement }> = ({ children }) => {
@@ -70,35 +53,12 @@ const PrivateRoute: React.FC<{ children: React.ReactElement }> = ({ children }) 
   // return isAuthenticated ? children : <Navigate to={ROUTES.LOGIN} replace />;
 };
 
-// 顧客関連ページ用のラッパーコンポーネント（動的にメニューを生成）
-const LessonHistoryWithMenu: React.FC = () => {
+// 顧客関連ページ用の共通レイアウトコンポーネント
+const CustomerRelatedLayout: React.FC<{ children: React.ReactNode }> = ({ children }) => {
   const { id } = useParams<{ id: string }>();
-  const menuItems = getLessonHistoryMenuItems(id || 'current');
-  return (
-    <MainLayout menuItems={menuItems}>
-      <LessonHistory />
-    </MainLayout>
-  );
-};
+  const menuItems = getLessonHistoryMenuItems(id ?? 'current');
 
-const CustomerProfileWithMenu: React.FC = () => {
-  const { id } = useParams<{ id: string }>();
-  const menuItems = getLessonHistoryMenuItems(id || 'current');
-  return (
-    <MainLayout menuItems={menuItems}>
-      <CustomerProfile />
-    </MainLayout>
-  );
-};
-
-const PostureListWithMenu: React.FC = () => {
-  const { id } = useParams<{ id: string }>();
-  const menuItems = getLessonHistoryMenuItems(id || 'current');
-  return (
-    <MainLayout menuItems={menuItems}>
-      <PostureList />
-    </MainLayout>
-  );
+  return <MainLayout menuItems={menuItems}>{children}</MainLayout>;
 };
 
 const AppRoutes: React.FC = () => {
@@ -107,16 +67,6 @@ const AppRoutes: React.FC = () => {
       <AuthProvider>
         <Routes>
           <Route path={ROUTES.LOGIN} element={<Login />} />
-          <Route
-            path={ROUTES.SHOP_MANAGEMENT}
-            element={
-              <PrivateRoute>
-                <MainLayout menuItems={shopManagementMenuItems}>
-                  <ShopManagement />
-                </MainLayout>
-              </PrivateRoute>
-            }
-          />
           <Route
             path={ROUTES.CUSTOMER_SELECT}
             element={
@@ -141,7 +91,9 @@ const AppRoutes: React.FC = () => {
             path={ROUTES.LESSON_HISTORY}
             element={
               <PrivateRoute>
-                <LessonHistoryWithMenu />
+                <CustomerRelatedLayout>
+                  <LessonHistory />
+                </CustomerRelatedLayout>
               </PrivateRoute>
             }
           />
@@ -149,7 +101,9 @@ const AppRoutes: React.FC = () => {
             path={ROUTES.CUSTOMER_PROFILE}
             element={
               <PrivateRoute>
-                <CustomerProfileWithMenu />
+                <CustomerRelatedLayout>
+                  <CustomerProfile />
+                </CustomerRelatedLayout>
               </PrivateRoute>
             }
           />
@@ -157,17 +111,9 @@ const AppRoutes: React.FC = () => {
             path={ROUTES.POSTURE_LIST}
             element={
               <PrivateRoute>
-                <PostureListWithMenu />
-              </PrivateRoute>
-            }
-          />
-          <Route
-            path={ROUTES.POSTURE_DETAIL}
-            element={
-              <PrivateRoute>
-                <MainLayout menuItems={defaultMenuItems}>
-                  <PostureDetail />
-                </MainLayout>
+                <CustomerRelatedLayout>
+                  <PostureList />
+                </CustomerRelatedLayout>
               </PrivateRoute>
             }
           />
@@ -221,7 +167,7 @@ const AppRoutes: React.FC = () => {
               </PrivateRoute>
             }
           />
-          <Route path="/" element={<Navigate to={ROUTES.SHOP_MANAGEMENT} replace />} />
+          <Route path="/" element={<Navigate to={ROUTES.CUSTOMER_SELECT} replace />} />
         </Routes>
       </AuthProvider>
     </BrowserRouter>
